@@ -1,7 +1,10 @@
-<div class="content">
-  <h1>Home</h1>
-  <p>This page is an example application of svelte with shoelace.</p>
-</div>
+<script lang="ts">
+  import Layout from "$src/layout/Layout.svelte";
+</script>
 <svelte:head>
   <title>Home</title>
 </svelte:head>
+<Layout>
+  <h1>Home</h1>
+  <p>This is an example application of <a href="https://svelte.dev/" target="_blank">svelte</a> with <a href="https://shoelace.style/" target="_blank">shoelace</a>.</p>
+</Layout>
