@@ -1,6 +1,9 @@
-<div class="content">
-  <h1>NotFound</h1>
-</div>
+<script lang="ts">
+  import Layout from "$src/layout/Layout.svelte";
+</script>
 <svelte:head>
   <title>Not Found</title>
 </svelte:head>
+<Layout>
+  <h1>Not Found</h1>
+</Layout>
