@@ -1,12 +1,12 @@
 <script>
-  import "@shoelace-style/shoelace/dist/components/progress-ring/progress-ring";
+  import { onMount } from "svelte";
 
-  import Layout from "$src/layout/Layout.svelte";
+  import "@shoelace-style/shoelace/dist/components/progress-ring/progress-ring.js";
 
-  let progress = $state(0);
-  let d = $state(0);
+  let progress = 0;
+  let d = 0;
 
-  $effect(() => {
+  onMount(() => {
     const timerID = setInterval(() => {
       d++;
       progress = Math.min(progress + d, 100);
@@ -17,12 +17,12 @@
     return () => clearInterval(timerID);
   });
 </script>
-<Layout>
+<div class="content">
   <h1>Progress Ring</h1>
   <div>
     <sl-progress-ring
       value={progress}
-      style="margin-bottom: 1.5em;"
+      style="margin-bottom: .5em;"
     >{progress}</sl-progress-ring>
   </div>
-</Layout>
+</div>
